update errorHandlerMiddleware file and rename other middlewares to be in same format
