--- src/middleware/erroHandlerMiddleware.js
@@ -1,10 +1,0 @@
-exports.errorHandler = (err, req, res, next) => {
-
-    const errStatus = err.statusCode || 500;
-    const errMsg = err.message || 'Something went wrong';
-    res.status(errStatus).json({
-        success: false,
-        status: errStatus,
-        message: errMsg,
-    })
-}
